+++ apps/api/src/app.module.ts
@@ -14,6 +14,7 @@ import { PlacesModule } from './places/places.module';
 import { SignalsModule } from './signals/signals.module';
 import { FairPlayModule } from './fair-play/fair-play.module';
 import { AdminConfigModule } from './admin-config/admin-config.module';
+import { ArtistBandsModule } from './artist-bands/artist-bands.module';
 
 @Module({
   imports: [
@@ -38,6 +39,7 @@ import { AdminConfigModule } from './admin-config/admin-config.module';
     SignalsModule,
     FairPlayModule,
     AdminConfigModule,
+    ArtistBandsModule,
   ],
 })
 export class AppModule {}